@@ -1,83 +1,92 @@
-declare namespace locatePath {
-	interface Options {
-		/**
-		Current working directory.
+export interface Options {
+	/**
+	The current working directory.
 
-		@default process.cwd()
-		*/
-		readonly cwd?: string;
+	@default process.cwd()
+	*/
+	readonly cwd?: string;
 
-		/**
-		Type of path to match.
+	/**
+	The type of path to match.
 
-		@default 'file'
-		*/
-		readonly type?: 'file' | 'directory';
+	@default 'file'
+	*/
+	readonly type?: 'file' | 'directory';
 
-		/**
-		Allow symbolic links to match if they point to the requested path type.
+	/**
+	Allow symbolic links to match if they point to the requested path type.
 
-		@default true
-		*/
-		readonly allowSymlinks?: boolean;
-	}
-
-	interface AsyncOptions extends Options {
-		/**
-		Number of concurrently pending promises. Minimum: `1`.
-
-		@default Infinity
-		*/
-		readonly concurrency?: number;
-
-		/**
-		Preserve `paths` order when searching.
-
-		Disable this to improve performance if you don't care about the order.
-
-		@default true
-		*/
-		readonly preserveOrder?: boolean;
-	}
+	@default true
+	*/
+	readonly allowSymlinks?: boolean;
 }
 
-declare const locatePath: {
+export interface AsyncOptions extends Options {
 	/**
-	Synchronously get the first path that exists on disk of multiple paths.
+	The number of concurrently pending promises.
 
-	@param paths - Paths to check.
-	@returns The first path that exists or `undefined` if none exists.
+	Minimum: `1`
+
+	@default Infinity
 	*/
-	sync: (
-		paths: Iterable<string>,
-		options?: locatePath.Options
-	) => string | undefined;
+	readonly concurrency?: number;
 
 	/**
-	Get the first path that exists on disk of multiple paths.
+	Preserve `paths` order when searching.
 
-	@param paths - Paths to check.
-	@returns The first path that exists or `undefined` if none exists.
+	Disable this to improve performance if you don't care about the order.
 
-	@example
-	```
-	import locatePath = require('locate-path');
-
-	const files = [
-		'unicorn.png',
-		'rainbow.png', // Only this one actually exists on disk
-		'pony.png'
-	];
-
-	(async () => {
-		console(await locatePath(files));
-		//=> 'rainbow'
-	})();
-	```
+	@default true
 	*/
-	(paths: Iterable<string>, options?: locatePath.AsyncOptions): Promise<
-	string | undefined
-	>;
-};
+	readonly preserveOrder?: boolean;
+}
 
-export = locatePath;
+/**
+Get the first path that exists on disk of multiple paths.
+
+@param paths - The paths to check.
+@returns The first path that exists or `undefined` if none exists.
+
+@example
+```
+import {locatePath} from 'locate-path';
+
+const files = [
+	'unicorn.png',
+	'rainbow.png', // Only this one actually exists on disk
+	'pony.png'
+];
+
+console(await locatePath(files));
+//=> 'rainbow'
+```
+*/
+export function locatePath(
+	paths: Iterable<string>,
+	options?: AsyncOptions
+): Promise<string | undefined>;
+
+/**
+Synchronously get the first path that exists on disk of multiple paths.
+
+@param paths - The paths to check.
+@returns The first path that exists or `undefined` if none exists.
+
+@example
+```
+import {locatePathSync} from 'locate-path';
+
+const files = [
+	'unicorn.png',
+	'rainbow.png', // Only this one actually exists on disk
+	'pony.png'
+];
+
+console(locatePathSync(files));
+//=> 'rainbow'
+```
+*/
+export function locatePathSync(
+	paths: Iterable<string>,
+	options?: Options
+): string | undefined;
